src/app/change-request/project/project-change-request-details/project-change-request-define-solution/project-change-request-define-scope-dialog/project-define-scope-dialog-info/project-define-scope-dialog-info.component.ts
import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';

@Component({
  selector: 'mc-project-define-scope-dialog-info',
  templateUrl: './project-define-scope-dialog-info.component.html',
  styleUrls: ['./project-define-scope-dialog-info.component.scss']
})
export class ProjectDefineScopeDialogInfoComponent implements OnInit {
  @Input() formConfiguration: any;
  @Output() readonly onConfirm: EventEmitter<void> = new EventEmitter<void>();
  @Output() readonly onCancel: EventEmitter<void> = new EventEmitter<void>();
  constructor() { }

  ngOnInit(): void {
  }

  onProceed() {
    this.onConfirm.emit();
  }

  onClose() {
    this.onCancel.emit();
  }

}
